refactor: make all pieces virtual
